receiptomatic-www: Fix the type of the session id (sid)
In the database, the sid (session id) is defined as an unsigned
bigint(20), this is the same as an unsigned long long and has a value
in the range, 0 - 18446744073709551615.

In the code we thus need to make sure we are using (unsigned long long)
for the sid.

We also need to make sure we are using 21 chars to hold the sid as a
string, 20 + NUL.

+++ www/common.h
@@ -84,7 +84,7 @@
  * in a tokyocabinet database table inbetween requests.
  */
 struct session {
-	unsigned int sid;
+	unsigned long long sid;
 	unsigned int uid;
 	unsigned char capabilities;
 	char *username;
